+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## HEAD
 
+### Public/Private Key Consistency
+
+ * #1044: If a public key has been configured for a log, check that it is consistent with the private key.
+ * #1046: Ensure that no two logs in the CTFE configuration use the same private key.
+
 ### Cleanup
 
  * Remove v2 log list package files.